Feat: update controller

+++ api/src/api/controllers/user.controller.ts
@@ -1,13 +1,12 @@
-import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Patch } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpCode, HttpStatus, Patch, UnauthorizedException } from "@nestjs/common";
 import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 
-import { ProfileResultDTO, UpdateNicknameDTO, UserModel, UserService, UserSessionModel, UserSessionService, WithdrawDTO } from "src/app/user";
+import { ProfileResultDTO, UpdateNicknameDTO, UpdatePasswordDTO, UserModel, UserService, UserSessionModel, UserSessionService, WithdrawDTO } from "src/app/user";
 
 import { Credential, Public } from "../decorators";
 
 @Controller({ path: 'users', version: '1' })
 @ApiTags('회원정보')
-@ApiBearerAuth()
 export class UserController {
 
     constructor(
@@ -16,6 +15,7 @@ export class UserController {
     ) { }
 
     @Get('me')
+    @ApiBearerAuth()
     @ApiOperation({ summary: '프로필 조회' })
     @ApiResponse({ status: HttpStatus.OK, type: ProfileResultDTO })
     getProfile(@Credential() user: UserModel): ProfileResultDTO {
@@ -23,6 +23,7 @@ export class UserController {
     }
 
     @Patch('nickname')
+    @ApiBearerAuth()
     @ApiOperation({ summary: '닉네임 변경' })
     @ApiResponse({ status: HttpStatus.OK, type: ProfileResultDTO })
     async updateNickname(@Credential() user: UserModel, @Body() dto: UpdateNicknameDTO): Promise<ProfileResultDTO> {
@@ -31,20 +32,28 @@ export class UserController {
     }
 
     @Public()
+    @HttpCode(HttpStatus.NO_CONTENT)
     @Patch('password')
-    @ApiOperation({ summary: '비밀번호 변경' })
-    @ApiResponse({ status: HttpStatus.OK, type: ProfileResultDTO })
-    async updatePassword(@Credential() user: UserModel, @Body() dto: UpdateNicknameDTO): Promise<ProfileResultDTO> {
-        const updatedUser = await this.userService.updateNickname(user, dto.nickname);
-        return ProfileResultDTO.from(updatedUser);
+    @ApiOperation({ summary: '비밀번호 변경 <OTP 연계>' })
+    @ApiResponse({ status: HttpStatus.NO_CONTENT })
+    async updatePassword(@Body() dto: UpdatePasswordDTO): Promise<void> {
+        const user = await this.userService.getUserByEmailOrThrow(dto.email);
+        if (!user.verifyOtp(dto.otp)) {
+            throw new UnauthorizedException('OTP 코드가 유효하지 않습니다.');
+        }
+        await this.userService.updatePassword(user, dto.password);
     }
 
     @Delete()
     @HttpCode(HttpStatus.NO_CONTENT)
+    @ApiBearerAuth()
     @ApiOperation({ summary: '회원탈퇴 <OTP 연계>' })
     @ApiResponse({ status: HttpStatus.NO_CONTENT })
     async withdraw(@Credential() user: UserModel, @Body() dto: WithdrawDTO): Promise<void> {
-        await this.userService.withdraw(user, dto, async () => {
+        if (!user.verifyOtp(dto.otp)) {
+            throw new UnauthorizedException('OTP 코드가 유효하지 않습니다.');
+        }
+        await this.userService.withdraw(user, async () => {
             const userSessions: UserSessionModel[] = await this.userSessionService.getUserSessions(user.id);
             await this.userSessionService.removes(userSessions);
         });
